requirements.txt
PyPDF2==3.0.1
PyMuPDF==1.23.8
tkinter-tooltip==2.1.0
Pillow==10.1.0
pyttsx3==2.90
playsound==1.3.0
Flask==2.3.3 